[ALPS05636285] mml: add racing mode
Add inline rotate (racing) mode support.

MTK-Commit-Id: fdf9f6fdd921c6a225a725b17024915248c0be1c

CR-Id: ALPS05636285
Feature: [Module]MML driver

+++ drivers/gpu/drm/mediatek/mml/mtk-mml-core.h
@@ -151,7 +151,7 @@ struct mml_topology_ops {
 	s32 (*init_cache)(struct mml_dev *mml,
 			  struct mml_topology_cache *cache,
 			  struct cmdq_client **clts,
-			  u8 clt_cnt);
+			  u32 clt_cnt);
 	s32 (*select)(struct mml_topology_cache *cache,
 		      struct mml_frame_config *cfg);
 };
